refactor: rename session-file-storage.ts to file-session-storage.ts
Match the FileSessionStorage class name and the <backend>-storage.ts
pattern of the bun stores. Update the two import paths.
