+++ src/node/index.ts
@@ -12,4 +12,4 @@ export * from "./debug.js";
 export * from "./from-path.js";
 export * from "./run.js";
 export * from "./server.js";
-export * from "./session-file-storage.js";
+export * from "./file-session-storage.js";
